+++ src/bytestring.h
@@ -51,7 +51,7 @@ public:
 	virtual void deserializeBinary(ByteStream &s);
 
 protected:
-	int readData(char *buffer, size_t size);
+	size_t readData(char *buffer, size_t size);
 	void writeData(const char *data, size_t size);
 };
