+++ src/socket.h
@@ -45,7 +45,7 @@ public:
 	void close(void);
 
 	// Stream, ByteStream
-	int readData(char *buffer, size_t size);
+	size_t readData(char *buffer, size_t size);
 	void writeData(const char *data, size_t size);
 
 private:
